bootstrap: create a new package for dev initialization
